layout/base/tests/test_bug332655-2.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=332655
-->
<head>
  <title>Test for Bug 332655</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/WindowSnapshot.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body onload="test()">
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=332655">Mozilla Bug 332655</a>
<p id="display"></p>
<div id="content">
<input type="text" id="testInput"
       style="-moz-appearance: none; appearance: none"> <!-- bug 1234659 workaround -->
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 332655 **/

SimpleTest.waitForExplicitFinish();

function repeatKey(key, repetitions) {
  for (var i = 0; i < 4; ++i) {
    synthesizeKey(key, {});
  }
}

function test() {
  var textInput = $("testInput");
  var s1, s2, s3, equal, str1, str2;

  textInput.focus();
  synthesizeKey("\u05d0", { });
  synthesizeKey("a", { });
  synthesizeKey("b", { });
  synthesizeKey(" ", { });
  synthesizeKey("\u05d1", { });
  synthesizeKey("\u05d2", { });
  s1 = snapshotWindow(window);

  // 4 LEFT to get to the beginning of the line: HOME doesn't work on OS X
  repeatKey("VK_LEFT", 4);
  synthesizeKey("VK_BACK_SPACE", { });
  synthesizeKey("\u05d0", { });
  s2 = snapshotWindow(window);

  [equal, str1, str2] = compareSnapshots(s1, s2, true);
  ok(equal, "deleting and inserting RTL char at beginning of line shouldn't change: expected " +
     str1 + " but got " + str2);

  textInput.select();
  synthesizeKey("a", { });
  synthesizeKey("b", { });
  synthesizeKey(" ", { });
  synthesizeKey("\u05d1", { });
  synthesizeKey("\u05d2", { });
  // 4 LEFT to get to the beginning of the line: HOME doesn't work on OS X
  repeatKey("VK_LEFT", 4);
  synthesizeKey("\u05d0", { });

  s3 = snapshotWindow(window);

  [equal, str1, str2] = compareSnapshots(s1, s3, true);
  ok(equal, "the order entering Bidi text shouldn't change rendering: expected " +
     str1 + " but got " + str2);

  SimpleTest.finish();
}

</script>
</pre>
</body>
</html>
